+++ src/plugin-api/main/uk/co/reecedunn/intellij/plugin/processor/query/QueryProcessor.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2020 Reece H. Dunn
+ * Copyright (C) 2018-2021 Reece H. Dunn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import java.io.Closeable
 interface QueryProcessor : Closeable {
     val presentation: ItemPresentation
 
-    val servers: List<String>
+    fun servers(database: String): List<String>
 
     val databases: List<String>
 }