+++ src/plugin-saxon/main/uk/co/reecedunn/intellij/plugin/saxon/query/s9api/SaxonQueryProcessor.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2020 Reece H. Dunn
+ * Copyright (C) 2018-2021 Reece H. Dunn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -73,7 +73,7 @@ internal class SaxonQueryProcessor(
             ItemPresentationImpl(SaxonIcons.Product, "Saxon ${processor.version}")
         }
 
-    override val servers: List<String> = listOf()
+    override fun servers(database: String): List<String> = listOf()
 
     override val databases: List<String> = listOf()