+++ CHANGELOG.md
@@ -25,6 +25,7 @@ XRay Unit Tests:
 Run Configuration:
 
 1. Support using the modules database for selected MarkLogic app-server.
+1. Filter the app-servers to the ones that match the selected database.
 
 ## 2020